🐛 fix(philo.h): change type of value in t_value struct from int to long to accommodate larger values 🐛 fix(philo.h): change type of times_to_eat in t_sim struct from long to int to match the type of other time variables 🐛 fix(philo.c): remove unused time_to_die_mutex and replace it with time_to_die_2 in philo_free function 🐛 fix(philo.c): replace time_to_die variable with time_to_die_2 in philo_eat function 🐛 fix(philo.c): remove time_to_die_mutex and replace it with time_to_die_2 in philo_create function 🐛 fix(sim.c): replace time_to_die variable with time_to_die_2 in someone_died function 🐛 fix(value.c): change type of value_get function from int to long to match the type of value in t_value struct 🐛 fix(value.c): change type of value_set function from int to long to match the type of value in t_value struct

+++ philo/src/tools/value.c
@@ -6,15 +6,15 @@
 /*   By: jensbouma <jensbouma@student.codam.nl>       +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/07/24 23:26:15 by jensbouma     #+#    #+#                 */
-/*   Updated: 2023/07/27 01:52:19 by jensbouma     ########   odam.nl         */
+/*   Updated: 2023/07/28 10:39:54 by jensbouma     ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-static int	value_get(t_value *v)
+static long	value_get(t_value *v)
 {
-	int	value;
+	long	value;
 
 	pthread_mutex_lock(&v->mutex);
 	value = v->value;
@@ -22,7 +22,7 @@ static int	value_get(t_value *v)
 	return (value);
 }
 
-static int	value_set(t_value *v, int value)
+static long	value_set(t_value *v, long value)
 {
 	pthread_mutex_lock(&v->mutex);
 	v->value = value;
